backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
Flask-RESTful==0.3.10
marshmallow==3.20.1